Extract logic from driver to allow usage as a library
